+++ src/dq_suite/schemas/profilingattribuut.py
@@ -3,7 +3,9 @@ from pyspark.sql.types import StructType
 SCHEMA = (
     StructType()
     .add("profilingAttribuutId", "string")
+    .add("profilingTabelId", "string")
     .add("bronAttribuutId", "string")
+    .add("attribuutNaam", "string")
     .add("missingDataPercentage", "double")
     .add("minWaarde", "string")
     .add("maxWaarde", "string")